Updated the way the Paychecks show on the profile
