+++ server/api/chips/chip_services.js
@@ -1,4 +1,4 @@
-import { determine_filter } from "../../util";
+import { determine_filter } from "../../utils/util";
 import { chip_db } from "../chips";
 
 export default {
