+++ server/api/users/user_db.js
@@ -1,5 +1,5 @@
 import User from "../users/user";
-import { prnt } from "../../util";
+import { prnt } from "../../utils/util";
 // import Token from "../tokens/token";
 require("dotenv");
