+++ server/config.js
@@ -2,9 +2,9 @@ import dotenv from "dotenv";
 
 dotenv.config();
 
-const environment = process.env.ENVIRONMENT;
+// const environment = process.env.ENVIRONMENT;
 // const environment = "staging";
-// const environment = "production";
+const environment = "production";
 
 const decideEnvironment = ({ production, staging, development }) => {
   if (environment === "development") {
