+++ server/email_templates/pages/current_stock.js
@@ -1,4 +1,4 @@
-import { format_date } from "../../util";
+import { format_date } from "../../utils/util";
 
 export default products => {
   const date = new Date();
